backend/requirements.txt
torch
transformers
datasets
pillow
pandas
fastapi
uvicorn
python-multipart
scikit-learn
accelerate
kagglehub
